.travis.yml: install git-buildpackage from Trusty
Ubuntu Trusty has git-buildpackage v0.6.9, so install that version
during the Travis CI tests. When Xenial is a stable/supported platform,
we can bump this version number to match Xenial's package.

Incidentally this v0.6.9 version of git-buildpackage is not compatible
with Python 3, so we only install it in the Python 2 environment.