build: workaround python-jenkins-0.2.1's build_job()
Prior to this change, the system version of Ubuntu Trusty's
python-jenkins did not properly submit new jobs to Jenkins. The problem
was that Jenkins expected a POST request, and python-jenkins-0.2.1 sent
a GET request instead.

To force python-jenkins to send a POST, copy upstream's solution
(https://bugs.launchpad.net/python-jenkins/+bug/1177831) and pass an
empty string to urllib2's Request constructor. The new
"_has_broken_build_job()" method checks to see if we have a buggy
python-jenkins version, and if not, we just fall through to the original
behavior (without monkeypatching).

This commit could be reverted if we ever require a newer Ubuntu OS
version (like Xenial) with a newer/less-buggy python-jenkins package.