merge-patches: always fetch from the patches remote
Prior to this change, "rhcephpkg merge-patches" would not always fetch
the latest patches remote. If the user was currently on a patch-queue
branch, rhcephpkg would run "git merge", which does not fetch the
remote.

If the user was on a debian branch instead, then
"rhcephpkg merge-patches" would fetch the remote as expected.

Make both situations do the same fetch+merge operation, so it's easier
to understand what this command does.