patch: do nothing if there are no new patches
Prior to this change, rhcephpkg patch was not idempotent. If you run it
twice with the same patch-queue branch, it will bump the release in
/debian/changelog each time, even if there are no changelog entries to
add.

It is better to exit without touching any file or generating any
dist-git commit.

Exit with a non-zero exit code so that my internal auto-build scripts
will not try to kick off any new Jenkins build. This is kind of a hack,
but it's the quickest solution for now.