tests: fix bare "localbuild" tambo subcommand
This test_localbuild() parameterized test checks that we will call
git-buildpackage with the proper "--git-dist=" argument, depending on
the command-line args that get passed through Tambo's arg parser.

The purpose of the test's first parameter is simply to verify that a
single subcommand "localbuild" without any args will result in a gbp
argument "--git-dist=trusty".

Unfortunately the first parameterized test was passing a simple string
"localbuild" to Tambo's arg parser, rather than a tuple of args like the
rest of the tests.

Tambo was calling split() on the string "localbuild" here, treating each
letter in "l", "o", "c", "a", "l" ... as a list member for
unknown_commands. In other words, Tambo was not recognizing the
"localbuild" subcommand in this case.

This does not cause any issue right now because we simply ignore
all unknown_commands in localbuild.py, but once we start sanity-checking
and bailing on unknown_commands, this test will fail.

Update the test parameters to use lists instead of the mistake I made
with the tuples here. This makes the code easier to read and verify that
we're passing the right iterable things through Tambo's parser.