merge-patches: support branch suffixes
Prior to this change, get_rhel_patches_branch() was translating a Debian
dist-git branch "ceph-2-ubuntu-hotfix-bz123" to "ceph-2-rhel-patches",
stripping the suffix.

The result was that merge-patches cannot fetch changes from a RHEL
-patches branch if it's a hotfix (or anything else special).

We need it to preserve the suffix in these cases, so merge-patches will
work with hotfix branches.