tests: call main() in test_patch.py
In regular operation, Tambo calls .main(), which populates the
self.parser variable and then calls _run(). When we call the private
_run() method directly, we circumvent the .parser definition.

Call the main() method so we exercise the code in a way that more
closely matches what a user does.

This also allows future-proofs us for when I have to add Tambo CLI
arguments to the "patch" subcommand.