patch: add generate_changelog() method
Move the changelog list construction code from the main _run() method
out into its own generate_changelog() method.

The purpose of this change is to reduce the size of _run() and make it
easier to re-use this code eventually in the non-simple case (where we
must read change information from "git status").