raise on bad package name
If we're not at the root of a Git clone, util.package_name() will return
whatever directory we happen to be in, rather than the name of the
package. If we've cd'd into the "debian" directory of a package, for
example, "rhcephpkg build" will tell Jenkins to to build a package named
"debian".

I think Fedora's rpkg tries to be a bit more clever here with parsing
the Git remote URL and so on, but that has its own problems. Let's
keep it simple for now and just pessimistically raise in this case.