build: custom Jenkins class for watching builds
python-jenkins has a bug where build_job() returns nothing,
https://launchpad.net/bugs/1724932

I've submitted changes to python-jenkins upstream to make it return a
queue ID number, which we can query with a new get_queue_item() method.

Until these changes are present in the python-jenkins package in the
Ubuntu distribution we support, add them to a child class
"RhcephpkgJenkins". Use this child class for builds.

This allows us to watch a build's queue ID and eventually pass the job
ID over to the "watch-build" code.