watch-build: add newline if necessary
If we watched an ongoing build, we'd print "<pkg> building for <time>"
lines to the console without newlines.

When we logged the "Ended" message, that would end up on the same line,
making it harder to read.

Track whether we're looping (writing messages without newlines) and add
a newline in that case.