rename "_patches" variables to "_patch_queue"
In some cases I was using variables and methods named "patches" to refer
to the debian "patch-queue/" branch.

This gets especially confusing in the "merge-patches" CLI sub-command,
where we have to interact with a "rhel patches branch" and "patch queue"
branch, and the difference between the two is very important.

Rename all the methods and variables that deal with git-buildpackage's
"patch-queue/" branches to have "patch_queue" names, rather than
the ambiguous "patches" names.