merge-patches: ensure local patch-queue branch exists
Prior to this change, the "merge-patches" command would always behave
like "--force" *if* no local patch-queue branch existed at all.

If the user is not on a local patch-queue branch already, ensure that
one at least exists and tracks origin.

With this change, "merge-patches" (without --force) will fail if the
merge from RHEL -> Debian is not a fast-forward merge, regardless of
which branch the user happens to have checked out at the moment.