tests: use testpkg fixture in localbuild test
The localbuild operation calls util.package_name(), and this method
expects to be at the root of a Git directory (as of
d833d56d732cfd7dc96d2b33a6d5de089f8d621d).

This means the tests fail when running outside of rhcephpkg.git (say, in
pbuilder with pybuild.)

Instead of monkeypatching further, use the testpkg fixture for this
test.

This means we have to tweak the expected behavior slightly (default
distro becomes "xenial" instead of "trusty"), because the debian branch
is ceph-2-ubuntu in that fixture, and get_distro() returns xenial for
that.