patch: fix old patch comparison on py3
In Python 3, map() returns a map object (iterator) instead of a list.

This meant that old_subjects was an iterator on py3, not a list. Later
on, when we checked new_series' subjects against old_subjects, we would
never find anything here:

   # Select only the ones that are new (according to commit subjects)
   new_series = [p for p in new_series if p.subject not in old_subjects]

Effectively, on py3, rhcephpkg would consider all edited patches to be
entirely "new", even when they were simply rebased, because this
old_subjects check would never filter out anything.

Use a list comprehension to construct old_subjects instead. This makes
the code easier to read and ensures old_subjects is a list on py2 and
py3.