patch: add RHBZ numbers to d/changelog for deleted patches
When we pass deleted files into gbp's Patch constuctor, .long_desc is a
blank string, because gbp cannot parse a file that does not exist.

This means that we had almost no data about the patches we were
deleting. In particular we were raising BZNotFound on every deleted
patch.

To work around this, read_deleted_patch_description() will load the
deleted file from Git with "git show", write that to a temporary file,
parse that temp file with gbp's Patch, and returns *that* .long_desc
instead for parsing.

Now that we no longer raise BzNotFound when deleting a patch, update the
test to verify that we have a good changelog entry now.