tests: use testpkg fixture in new-version tests
The tmpgitdir fixture here was a tiny bit faster than the full "testpkg"
fixture, but the difference was only a few microseconds.

It's not worth maintaining this fixture here for two reasons:

1) It's more complex test code to maintain

2) The fixture is not even representative of a Git clone where we would
run "rhcephpkg new-version". It's getting to the point where the lack of
accuracy hinders bug fixes and features.

Drop the "tmpgitdir" fixture and use the full "testpkg" fixture in the
new-version tests.