watch-build: keyboard interrupt exits with non-zero
Prior to this change, if I was watching a build and ctrl-c'd out of it,
rhcephpkg would exit with a zero exit code

That meant that if I had the following shell command set up:

  rhcephpkg watch-build 482 && git push

... and then ctrl-c'd out of "watch-build", the "git push" would still
happen, too early than what I expected.

Make ctrl-c cause "watch-build" to exit with a non-zero exit code so any
chained commands with "&&" will not proceed.