setup.py: sleep before pushing to master
GitHub does not permit me to push to master if Travis CI has not marked
commits as "built".

Prior to this change, "git push origin master" would fail for the
version bump commit.

Push the tag, then sleep for Travis CI to build the tag, then continue
on pushing to master.

This is racy, but I'm just trying to incrementally improve this for now.